Generate a unique ID for each contact.
Rather than use `contact.recordID`, generate a `helloAgainID` for
each contact during contact import, consisting of the contact's given
and family names lowercased, and joined together with a hash of the
recordID using dashes.

This unique ID format can later be used as a reasonable looking deep
link URL to the contact, and it creates a persistent way of identifying
a user.

I'm doing this hashing rather than generating a random string so that
unit tests get consistent results. I'm using the `string-hash` library
and it seems fast enough.

I'm only using 16 bits of the hash, because I want the ID to be as short
as possible, and a birthday attack analysis suggests that you'd need to
have 37 contacts with the *exact same name* to get a 1% chance of
collision. Seems pretty unlikely.