Store unique IDs back to the native Contacts store.
In point of fact, store the unique ID as a URL address field of type
"HelloAgain" in the corresponding native Contact record. This custom URL
will actually be visible to the user, and be _clickable_ once we get
deep links into the app wired up in iOS.

The deep link takes the form: helloagain://to/first-last-abcd

The code currently writes back the unique deep link whenever a
friend is toggled. Also, when contacts are imported, they're searched
for missing deep links and the links are written back at that time,
although this shouldn't really happen.

What's great about this is that now any aspect of a native contact
record can change (aside from the HelloAgain deep link) and the
contact's identity will still match up with their HelloAgain friend
record -- even across devices (or even possibly platforms, when we get
that far).