Replace redux-thunk with redux-side-effect.
Hopefully this will simplify matters, make testing easier.

Now the side effect of storing back the helloAgainID on toggleContact is
done triggered in the friends reducer, in a way which is possible to
isolate, and keep the reducer as pure functional for testing.

Also, fortunately in practice, the side effect is idempotent.